initial excel-export DI prep
