+++ scripts/test_di_integration.py
